@Aurelienvd 2022-08-24
13.0 PR: https://github.com/odoo/odoo/pull/98844